+++ src/components/AppBar/index.js
@@ -1,4 +1,3 @@
-export { default as ActionButton } from './ActionButton/ActionButton';
 export { default as AppBar } from './AppBar/AppBar';
 export { default as BackAppBar } from './BackAppBar/BackAppBar';
 export { default as CloseAppBar } from './CloseAppBar/CloseAppBar';
