+++ src/components/ServerErrorFallback/index.js
@@ -1,2 +1,1 @@
 export { default as ServerErrorFallback } from './ServerErrorFallback';
-export { default as ResetButton } from './ResetButton/ResetButton';
