+++ src/components/index.js
@@ -35,6 +35,4 @@ export * from './ServerErrorFallback';
 export * from './ShimmerEffect';
 export * from './Sidebar';
 export * from './StageDots';
-export * from './Submit';
 export * from './Toast';
-export * from './WriteButton';
