+++ src/pages/ExamReviewEditPage/ExamReviewEditPage.jsx
@@ -6,7 +6,8 @@ import { editReviewDetail } from '@/apis';
 
 import { useToast } from '@/hooks';
 
-import { ActionButton, CloseAppBar } from '@/components/AppBar';
+import { ActionButton } from '@/shared/component';
+import { CloseAppBar } from '@/components/AppBar';
 import {
   CategoryButton,
   CategoryFieldset,
