+++ src/pages/ExamReviewsPage/ExamReviewsPage.jsx
@@ -4,14 +4,8 @@ import { useQuery } from '@tanstack/react-query';
 import { getNoticeLine } from '@/apis';
 import { useScrollRestoration } from '@/hooks';
 import { SearchExamReviewsSuspense } from '@/pages/ExamReviewsPage';
-import {
-  AppBar,
-  Search,
-  WriteButton,
-  Icon,
-  Filter,
-  Filters,
-} from '@/components';
+import { WriteButton } from '@/shared/component';
+import { AppBar, Search, Icon, Filter, Filters } from '@/components';
 
 import {
   YEARS,
