+++ src/pages/LoginPage/SignUpPage/SignUpPageStages/AccountInfoPage.jsx
@@ -1,7 +1,7 @@
 import { useState } from 'react';
 
 import { Input, SpecialInput } from '@/components/Input';
-import { Submit } from '@/components/Submit';
+import { Submit } from '@/shared/component';
 
 import {
   checkName,
