+++ src/pages/MyPage/pages/ChangePasswordPage/ChangePasswordPage.jsx
@@ -6,7 +6,8 @@ import { updatePassword } from '@/apis';
 
 import { useToast } from '@/hooks';
 
-import { BackAppBar, ActionButton, InputPassword } from '@/components';
+import { ActionButton } from '@/shared/component';
+import { BackAppBar, InputPassword } from '@/components';
 
 import { MUTATION_KEY, TOAST } from '@/constants';
