+++ src/pages/PostsPage/PostsPage.jsx
@@ -3,7 +3,9 @@ import { useQuery } from '@tanstack/react-query';
 
 import { getNoticeLine } from '@/apis';
 import { useScrollRestoration } from '@/hooks';
-import { BackAppBar, Icon, WriteButton } from '@/components';
+
+import { WriteButton } from '@/shared/component';
+import { BackAppBar, Icon } from '@/components';
 import { PostsSuspense } from '@/pages/PostsPage';
 import { getBoard } from '@/utils';
 import { QUERY_KEY } from '@/constants';
