+++ src/shared/component/index.js
@@ -1,0 +1,4 @@
+export { default as ActionButton } from './button/ActionButton/ActionButton';
+export { default as ResetButton } from './button/ResetButton/ResetButton';
+export { default as Submit } from './button/Submit/Submit';
+export { default as WriteButton } from './button/WriteButton/WriteButton';
